piece.py file edit to fix a bug in the check/checkmate logic
I think this bug is the same one that Heather was talking about, but can explain this over huddle if it's not clear. So, if a king was in check and could be blocked by a piece to get it out of check, it wasn't allowing you to move the piece to block the check. But, now this has been fixed.